+++ packages/app/utils/translation.ts
@@ -1,5 +1,51 @@
 import i18n, { TranslateOptions } from 'i18n-js'
 
+interface Language {
+  langCode: string
+  languageName: string
+  languageLocalName: string
+  active: boolean
+}
+
+export const languages: Language[] = [
+  {
+    langCode: 'sv',
+    languageName: 'Swedish',
+    languageLocalName: 'Svenska',
+    active: true,
+  },
+  {
+    langCode: 'en',
+    languageName: 'English',
+    languageLocalName: 'English',
+    active: true,
+  },
+  {
+    langCode: 'pl',
+    languageName: 'Polish',
+    languageLocalName: 'Polski',
+    active: true,
+  },
+  {
+    langCode: 'de',
+    languageName: 'German',
+    languageLocalName: 'Deutsch',
+    active: true,
+  },
+  {
+    langCode: 'ar',
+    languageName: 'Arabic',
+    languageLocalName: 'اَلْعَرَبِيَّةُ',
+    active: false,
+  },
+  {
+    langCode: 'so',
+    languageName: 'Somali',
+    languageLocalName: 'af-Soomaali',
+    active: false,
+  },
+]
+
 export const translations = {
   ar: require('../translations/ar.json'),
   de: require('../translations/de.json'),
